+++ integrations/gitlab/CHANGELOG.md
@@ -14,6 +14,54 @@ this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - Fixed glob pattern matching by utilizing GitLab's glob path search capabilities instead of pattern matching.
 
 
+0.2.31 (2025-02-04)
+===================
+
+### Improvements
+
+- Bumped ocean version to ^0.18.8
+
+
+0.2.30 (2025-01-29)
+===================
+
+### Improvements
+
+- Bumped ocean version to ^0.18.6
+
+
+0.2.29 (2025-01-28)
+===================
+
+### Improvements
+
+- Bumped ocean version to ^0.18.5
+
+
+0.2.28 (2025-01-27)
+===================
+
+### Bug Fixes
+
+- Revert changes made in 0.2.1
+
+
+0.2.27 (2025-01-23)
+===================
+
+### Improvements
+
+- Bumped ocean version to ^0.18.4
+
+
+0.2.26 (2025-01-22)
+===================
+
+### Improvements
+
+- Bumped ocean version to ^0.18.3
+
+
 0.2.25 (2025-01-22)
 ===================